requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
Django
django-allauth==0.61.1
ecdsa==0.18.0
idna==3.6
kombu==5.3.6
oauthlib==3.2.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stripe==8.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
